backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==11.1.0
gunicorn==23.0.0